+++ core/line_runereader.go
@@ -1,3 +1,5 @@
+// +build !plan9
+
 package core
 
 import (
